Split off parser tests
